refactor: splice out product page functions
